split up views from AC class core
