+++ autocomplete/__init__.py
@@ -2,6 +2,7 @@
 Allow HTMXAutoComplete to be imported from the module directly
 """
 
-from .autocomplete import Autocomplete, register, urls
+from .core import Autocomplete, register
 from .shortcuts import ModelAutocomplete
+from .views import urls
 from .widgets import AutocompleteWidget
